requirements.txt
flask
flask-cors
tensorflow
numpy
pillow
scikit-learn
matplotlib
reportlab
